+++ test/sql/viewer.sql
@@ -40,6 +40,7 @@ select count(*) from emaj_mySchema1.myTbl1_log;
 -----------------------------
 -- authorized functions
 -----------------------------
+select * from emaj.emaj_get_current_log_table('myschema1', 'mytbl1');
 select * from emaj.emaj_verify_all();
 select emaj.emaj_get_previous_mark_group('myGroup1', current_timestamp);
 select emaj.emaj_get_previous_mark_group('myGroup1', 'EMAJ_LAST_MARK');